refactor: move rule calculations out of the sheet layer
docs/architecture.md の課題「プレゼンテーション層にルール計算」。
applications/helpers.ts が3つの役割を抱えていたので分割した。

- ルール計算 → module/rules/character-points.ts
  引数を CharacteristicsMap から Record<string, {value:number}> に絞り、
  applications への依存をなくした。テストもそのまま移動している
- DOMユーティリティ → module/utils/sheet.ts
  Foundryのシート汎用でエモクロア固有ではない
- 選択肢の組み立て → applications/helpers.ts に残す

外部からの呼び出し口は calculateCharPointSum と calculateTotalSkillPoints の
2箇所だけだったので影響範囲は小さい。

暗黙になっていたルールをコメントで明示した。運勢を能力値ポイントから
除外していること、エクストラ技能が skills と exSkills の両方で数えられて
コストが2倍になること。

あわせてデッドコードを削除: _getBaseSkills（呼び出し元なし）、
SkillPointCalculation（未使用）。
